pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-project-template"
version = "0.1.0"
description = "A versatile, production-ready Django project template"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Nodir ustoz", email = "nodir@example.com" }
]
keywords = ["django", "template", "copier", "saas", "production"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

# No runtime dependencies - this is a template repository
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.5.0",
    "pytest-timeout>=2.2.0",
    "copier>=9.0.0",
    "pyyaml>=6.0.0",
    "types-PyYAML>=6.0.0",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
    "plumbum>=1.8.0",
    "pre-commit>=3.8.0",
]

[project.urls]
Homepage = "https://github.com/NodirUstoz/Django-project-template"
Documentation = "https://github.com/NodirUstoz/Django-project-template"
Repository = "https://github.com/NodirUstoz/Django-project-template"
Issues = "https://github.com/NodirUstoz/Django-project-template/issues"

[tool.setuptools]
# This is a template repository, not a Python package
# We only need the tests/ directory for the test suite
packages = []
py-modules = []

[tool.ruff]
# Target Python 3.12+
target-version = "py312"
line-length = 100

[tool.ruff.lint]
# Enable comprehensive rule sets
select = [
    "F",     # Pyflakes
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "T10",   # flake8-debugger
    "RUF",   # Ruff-specific rules
]

ignore = []

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
# Test files can use fixtures and have imports that appear unused
"tests/*" = ["F811"]
# Django migrations are auto-generated and don't need to follow all rules
"*/migrations/*" = ["E501", "RUF012"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "-p", "no:cacheprovider",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests that require external dependencies",
    "generation: marks tests that generate projects",
    "feature: marks tests for specific features",
    "django: marks tests that interact with Django",
]
